+++ components/merkledb/src/proof_map_index/mod.rs
@@ -36,7 +36,6 @@ use crate::{
     },
     BinaryKey, BinaryValue, HashTag, ObjectHash,
 };
-use crate::validation::{is_valid_artifact_name, is_valid_service_name};
 
 mod key;
 mod node;