+++ CHANGELOG.md
@@ -34,6 +34,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
   - Removed obsolete `TestKit::blockchain_mut` method and `TestKit::blockchain`
   now returns value instead of reference.
+  
+- Dot symbol is not allowed in service names. (#1558)
 
 - Placeholder for changes due to dynamic services (#9999)
 
@@ -92,6 +94,8 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
   - For `Hash` keys both map variants works the same, because `ObjectHash`
   implementation for `Hash` returns the hash itself.
+  
+- Dot symbol is not allowed in index and group prefixes. (#1558)
 
 ### Internal improvements